Enum postgres_shared::types::Date

pub enum Date<T> {
    PosInfinity,
    NegInfinity,
    Value(T),
}

A wrapper that can be used to represent infinity with Type::Date types.

Variants
PosInfinity

Represents infinity, a date that is later than all other dates.

NegInfinity

Represents -infinity, a date that is earlier than all other dates.

Value(T)

The wrapped date.